backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-Cors==4.0.0
pytesseract==0.3.10
Pillow==10.0.0
mysql-connector-python==8.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
eventlet==0.33.3